Added login code files
There seems to be a limit of 100 files for an upload, so these are the just the files that contain the code that I wrote. I've included a picture of what the login screen looks like on my computer. The flutter code doesn't check for the username at all, and the password is password. Very secure